Split archive/remove alex tests
