+++ packages/types/src/Metadata/v2/Storage.ts
@@ -118,9 +118,7 @@ export class StorageFunctionType extends EnumType<PlainType | MapType> {
   toString (): string {
     if (this.isDoubleMap) {
       return `DoubleMap<${this.asDoubleMap.toString()}>`;
-    }
-
-    if (this.isMap) {
+    } else if (this.isMap) {
       if (this.asMap.isLinked) {
         return `(${this.asMap.value.toString()}, Linkage<${this.asMap.key.toString()}>)`;
       }
